refactor course details to exlude data retrieval
