+++ shared/GitVersion.txt
@@ -1,1 +1,1 @@
-0.1.0
+0.2.0
