+++ shared/ViewComponents/CourseDetailsViewComponent.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Linq;
 using GovUk.Education.SearchAndCompare.Domain.Client;
-using GovUk.Education.SearchAndCompare.UI.Shared.Services;
 using GovUk.Education.SearchAndCompare.UI.Shared.ViewModels;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ApplicationParts;
@@ -10,30 +9,9 @@ namespace GovUk.Education.SearchAndCompare.UI.Shared.ViewComponents
 {
     public class CourseDetailsViewComponent : ViewComponent
     {     
-
-        private readonly ICourseDetailsService _api;
-
-        public CourseDetailsViewComponent(ICourseDetailsService api)
+        public IViewComponentResult Invoke(CourseDetailsViewModel course)
         {
-            _api = api;
-        }
-
-        public IViewComponentResult Invoke(string providerCode, string courseCode, bool preview = false, string aboutYourOrgLink = null)
-        {
-            var course = _api.GetCourse(providerCode, courseCode);
-            var feeCaps = _api.GetFeeCaps();
-
-            var latestFeeCaps = feeCaps.OrderByDescending(x => x.EndYear).FirstOrDefault();
-
-            var viewModel = new CourseDetailsViewModel()
-            {
-                Course = course,
-                Finance = new FinanceViewModel(course, latestFeeCaps),
-                PreviewMode = preview,
-                AboutYourOrgLink = aboutYourOrgLink
-            };
-
-            return View(viewModel);
+            return View(course);
         }
     }
 }
