+++ src/Controllers/CourseController.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 using GovUk.Education.SearchAndCompare.UI.ActionFilters;
 using System.Linq;
 using GovUk.Education.SearchAndCompare.UI.Filters;
+using GovUk.Education.SearchAndCompare.UI.Shared.ViewModels;
 
 namespace GovUk.Education.SearchAndCompare.UI.Controllers
 {
@@ -13,13 +14,29 @@ namespace GovUk.Education.SearchAndCompare.UI.Controllers
     //[Authorize]
     public class CourseController : CommonAttributesControllerBase
     {
+        private readonly ISearchAndCompareApi _api;
+
+        public CourseController(ISearchAndCompareApi api)
+        {
+            _api = api;
+        }
+
         [HttpGet("course/{providerCode}/{courseCode}", Name = "Course")]
         public IActionResult Index(string providerCode, string courseCode, ResultsFilter filter)
         {
-            return View(new CourseViewModel {
-                CourseCode = courseCode,
-                ProviderCode = providerCode
-            });
+            var course = _api.GetCourse(providerCode, courseCode);
+            var feeCaps = _api.GetFeeCaps();
+
+            var latestFeeCaps = feeCaps.OrderByDescending(x => x.EndYear).FirstOrDefault();
+
+            var viewModel = new CourseDetailsViewModel()
+            {
+                Course = course,
+                Finance = new Shared.ViewModels.FinanceViewModel(course, latestFeeCaps),
+                PreviewMode = false
+            };
+
+            return View(viewModel);
         }
     }
 }
