--- src/ViewModels/CourseViewModel.cs
@@ -1,8 +1,0 @@
-namespace GovUk.Education.SearchAndCompare.UI.ViewModels
-{
-    public class CourseViewModel
-    {
-        public string CourseCode { get; set; }
-        public string ProviderCode { get; set; }        
-    }
-}
