+++ src/Views/Course/Index.cshtml
@@ -1,3 +1,3 @@
-@model CourseViewModel
+@model GovUk.Education.SearchAndCompare.UI.Shared.ViewModels.CourseDetailsViewModel
 
-@await Component.InvokeAsync("CourseDetails", new { providerCode = Model.ProviderCode, courseCode = Model.CourseCode });
+@await Component.InvokeAsync("CourseDetails", new { course = Model });
